refactor: :recycle: refactor servicebus-config
This commit mainly refactors the servicebus-configuration.sh file
by consolidating functions to reduce lines in main and tie the
existence checks directly to creating resources.

We also finished adding the queue creation resource functions.
Doing this in a new subscription lead to the provider error, so
provider registration was added.